1,不允許編輯表單給 user 權限使用，但允許 admin或manager使用 2,增加 can 的middleware，讓新增表單、儲存資料、刪除資料路由只能讓admin使用 3,讓每一個路由都必須經過 auth 的 middleware (即必須登入帳號才可以使用)。 4,更新兩個views： players.index 和teams.index，讓他們針對admin 與 manager 呈現受控管的超連結(新增、修改、刪除) 5,自動生成三個 users: admin, manager, and user (方便我們做測試之用) 6, 編輯 app/User.php，加入三種帳號角色名稱的常數字串，並將 role 這個欄位加入 fillable 中。 7, 編輯 database/migrations/2014_10_12_000000_create_users_table.php 這一個建立使用者資料表的 migration 設定檔，在其中加入一欄用於儲存帳號角色的 role 欄位。 8, 編輯 app/Http/Controllers/Auth/RegisterController.php，設定使用者註冊時的角色都是一般使用者。 9, 建立 Gate 權限規則。

+++ routes/web.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\PartiesController;
 use App\Http\Controllers\TitlesController;
+use Illuminate\Support\Facades\Auth;
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -32,20 +33,25 @@ Route::get('parties/green', [PartiesController::class, 'green'])->name('parties.
 Route::get('parties', [PartiesController::class, 'index'])->name('parties.index');
 Route::get('parties/{id}', [PartiesController::class, 'show'])->where('id', '[0-9]+')->name('parties.show');
 Route::get('parties/{id}/edit', [PartiesController::class, 'edit'])->where('id', '[0-9]+')->name('parties.edit');
-Route::delete('parties/delete/{id}', [PartiesController::class, 'destroy'])->where('id', '[0-9]+')->name('parties.destroy');
-Route::get('parties/create', [PartiesController::class, 'create'])->name('parties.create');
+Route::delete('parties/delete/{id}', [PartiesController::class, 'destroy'])->where('id', '[0-9]+')->name('parties.destroy')->middleware('can:admin');
+Route::get('parties/create', [PartiesController::class, 'create'])->name('parties.create')->middleware('can:admin');
 Route::get('parties/{id}/edit', [PartiesController::class, 'edit'])->where('id', '[0-9]+')->name('parties.edit');
 Route::patch('parties/update/{id}', [PartiesController::class, 'update'])->where('id', '[0-9]+')->name('parties.update');
-Route::post('parties/store', [PartiesController::class, 'store'])->where('id', '[0-9]+')->name('parties.store');
+Route::post('parties/store', [PartiesController::class, 'store'])->where('id', '[0-9]+')->name('parties.store')->middleware('can:admin');
 // 顯示顯示男代表資料
 Route::get('titles/session', [TitlesController::class, 'session'])->name('titles.session');
-Route::post('titles/city', [TitlesController::class, 'city'])->name('titles.city');
+Route::get('titles/city', [TitlesController::class, 'city'])->name('titles.city');
 
 Route::get('titles', [TitlesController::class, 'index'])->name('titles.index');
 Route::get('titles/{id}', [TitlesController::class, 'show'])->where('id', '[0-9]+')->name('titles.show');
 Route::get('titles/{id}/edit', [TitlesController::class, 'edit'])->where('id', '[0-9]+')->name('titles.edit');
-Route::delete('titles/delete/{id}', [TitlesController::class, 'destroy'])->where('id', '[0-9]+')->name('titles.destroy');
-Route::get('titles/create', [TitlesController::class, 'create'])->name('titles.create');
+Route::delete('titles/delete/{id}', [TitlesController::class, 'destroy'])->where('id', '[0-9]+')->name('titles.destroy')->middleware('can:admin');
+Route::get('titles/create', [TitlesController::class, 'create'])->name('titles.create')->middleware('can:admin');
 Route::get('titles/{id}/edit', [TitlesController::class, 'edit'])->where('id', '[0-9]+')->name('titles.edit');
 Route::patch('titles/update/{id}', [TitlesController::class, 'update'])->where('id', '[0-9]+')->name('titles.update');
-Route::post('titles/store', [TitlesController::class, 'store'])->where('id', '[0-9]+')->name('titles.store');
+Route::post('titles/store', [TitlesController::class, 'store'])->where('id', '[0-9]+')->name('titles.store')->middleware('can:admin');
+
+
+
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
+Auth::routes();
